+++ src/Audit.DynamicProxy/InterceptionSettings.cs
@@ -40,5 +40,9 @@ namespace Audit.DynamicProxy
         /// Gets or sets the event creation policy to use for this interception. Default is NULL to use the globally configured creation policy.
         /// </summary>
         public EventCreationPolicy? EventCreationPolicy { get; set; }
+        /// <summary>
+        /// Gets or sets the custom audit scope factory. Default is NULL to use the general AuditScopeFactory.
+        /// </summary>
+        public IAuditScopeFactory AuditScopeFactory  { get; set; }
     }
 }